Added galaxies_select_bootes_full.ipynb
This is an OLD version of the notebook used to select the sample of galaxies and classify them as quiescent or star-forming based on their NUV-R vs R-J colors, as well as to  create cutouts for each source from IR and radio maps and to stack them within different redshift and stellar mass bins. This old version includes some early pieces of code that I hope to delete soon, included here for the sake of safekeeping.